move zobrist.h from utils/ to core/
Bench:
